+++ ml/train/loop.go
@@ -132,6 +132,21 @@ func (loop *Loop) step(spec any, inputs, labels []*tensors.Tensor) (metrics []*t
 		return nil, err
 	}
 
+	// Free inputs and labels:
+	for _, input := range inputs {
+		input.FinalizeAll()
+	}
+	for _, label := range labels {
+		label.FinalizeAll()
+	}
+
+	// Free metrics on-device usage: on-device memory being more at premium,
+	// we want to immediately free things that are no longer used there.
+	for _, m := range metrics {
+		m.MaterializeLocal()
+		m.InvalidateOnDevice()
+	}
+
 	// Call "OnStep" hooks.
 	loop.onStep.Enumerate(func(hook *hookWithName[OnStepFn]) {
 		if err != nil {
@@ -178,6 +193,8 @@ func (loop *Loop) end(metrics []*tensors.Tensor) (err error) {
 // RunSteps runs those many steps. StartStep and EndStep are adjusted to the current
 // LoopStep, so it can be called multiple times, and it will simply pick up
 // where it left of last time.
+//
+// Note: inputs and labels yielded by the dataset are immediately finalized (freed) after use in each step.
 func (loop *Loop) RunSteps(ds Dataset, steps int) (metrics []*tensors.Tensor, err error) {
 	if steps == 0 {
 		return nil, nil
@@ -203,10 +220,20 @@ func (loop *Loop) RunSteps(ds Dataset, steps int) (metrics []*tensors.Tensor, er
 			}
 			return nil, errors.WithMessagef(err, "Loop.RunSteps(%d): failed reading from Dataset", steps)
 		}
+
+		// Immediately free any space being used.
+		for _, metric := range metrics {
+			metric.FinalizeAll()
+		}
 		metrics, err = loop.step(spec, inputs, labels)
 		if err != nil {
 			return nil, errors.WithMessagef(err, "Loop.RunSteps(%d): failed TrainStep(LoopStep=%d)", steps, loop.LoopStep)
 		}
+	}
+	for _, metric := range metrics {
+		// Transfer results locally and immediately free on-device storage.
+		metric.MaterializeLocal()
+		metric.InvalidateOnDevice()
 	}
 	err = loop.end(metrics)
 	if err != nil {
@@ -222,6 +249,8 @@ func (loop *Loop) RunSteps(ds Dataset, steps int) (metrics []*tensors.Tensor, er
 // be adjusted to expectation after the first epoch, when one knows how many steps there are
 // going to be.
 // Dataset.Reset is called after each epoch (including the last).
+//
+// Note: inputs and labels yielded by the dataset are immediately finalized (freed) after use in each step.
 func (loop *Loop) RunEpochs(ds Dataset, epochs int) (metrics []*tensors.Tensor, err error) {
 	if err = loop.Trainer.ResetTrainMetrics(); err != nil {
 		return